requirements.txt
# ============================================================================
# Core LangChain & LangGraph Dependencies
# ============================================================================
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0

# ============================================================================
# Web Framework & API
# ============================================================================
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
streamlit>=1.40.0

# ============================================================================
# Data Processing & Validation
# ============================================================================
pydantic>=2.0.0
pyyaml>=6.0.0

# ============================================================================
# File Processing
# ============================================================================
python-docx>=1.1.0
pdfplumber>=0.11.0

# ============================================================================
# Visualization
# ============================================================================
matplotlib>=3.8.0

# ============================================================================
# HTTP & Networking
# ============================================================================
requests>=2.32.0
httpx>=0.27.0

# ============================================================================
# Database (SQLite is built-in, no extra package needed)
# ============================================================================

# ============================================================================
# Optional: Development & Testing
# ============================================================================
# pytest>=8.0.0
# pytest-asyncio>=0.24.0
